server/requirements.txt
flask==2.0.3
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
Werkzeug==2.0.3
WTForms==2.2.1
